# backend_central_auth/swagger.yaml
openapi: 3.0.0
info:
  title: AeroGuard ZTNA — Central Auth
  description: Identity management, login, vendor provisioning, and dashboard APIs for Zero Trust Network Access
  version: 1.0.0
  contact:
    name: AeroGuard Team
    url: https://github.com/Sithumrulzs/backend_central_auth
servers:
  - url: http://localhost:8001
    description: Local development server
  - url: https://69e1efef-e429-472f-bfce-68e0ac0360ff-dev.e1-us-east-azure.choreoapis.dev/default/backendcentralauth/v1.0
    description: Choreo production (WSO2 Cloud)
...
paths:
...
  /api/v1/provision-vendor:
    post:
      tags:
        - Vendor Management
      summary: Provision vendor JIT access
      x-wso2-disable-security: true
      description: |
        Create a new vendor session with time-bound access.
        Generates a QR token for scanning at the gateway.
        Vendor credentials are ephemeral (no login password).
      operationId: provision_vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorProvisionPayload"
      responses:
        "200":
          description: Vendor provisioned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: profile_synced
                  message:
                    type: string
                    example: Vendor session created.
                  generated_qr_string:
                    type: string
                    example: AEROGUARD_QR_ACMECORP_1717504800
        "500":
          description: Provisioning failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Provision failed - database error
...
components:
  schemas:
...
    VendorProvisionPayload:
      type: object
      required:
        - vendor_username
        - company_name
        - clearance_level
        - target_device_id
        - valid_until
      properties:
        vendor_username:
          type: string
          description: Vendor account identifier
          example: vendor_acme
        company_name:
          type: string
          description: Vendor company name
          example: ACME Corp
        clearance_level:
          type: string
          description: Access clearance tier
          enum: [basic, standard, premium]
          example: standard
        target_device_id:
          type: string
          description: Target device or laptop identifier
          example: laptop_vendor_001
        valid_until:
          type: string
          format: date-time
          description: ISO8601 expiration timestamp
          example: "2026-06-11T23:59:59Z"
        qr_token:
          type: string
          description: Optional custom QR token (auto-generated if omitted)
          example: AEROGUARD_QR_CUSTOM_TOKEN_12345
...